+++ src/main/java/com/microsoft/graph/models/ExtensionSchemaProperty.java
@@ -48,7 +48,7 @@ public class ExtensionSchemaProperty implements IJsonBackedObject {
 
     /**
      * The Type.
-     * The type of the property that is defined as part of a schema extension.  Allowed values are Binary, Boolean, DateTime, Integer or String.  See the table below for more details.
+     * The type of the property that is defined as part of a schema extension.  Allowed values are Binary, Boolean, DateTime, Integer or String. See the table below for more details.
      */
     @SerializedName(value = "type", alternate = {"Type"})
     @Expose