+++ src/main/java/com/microsoft/graph/models/SchemaExtension.java
@@ -56,7 +56,7 @@ public class SchemaExtension extends Entity implements IJsonBackedObject {
 
     /**
      * The Status.
-     * The lifecycle state of the schema extension. Possible states are InDevelopment, Available, and Deprecated. Automatically set to InDevelopment on creation. Schema extensions provides more information on the possible state transitions and behaviors. Supports $filter (eq).
+     * The lifecycle state of the schema extension. Possible states are InDevelopment, Available, and Deprecated. Automatically set to InDevelopment on creation. For more information about the possible state transitions and behaviors, see Schema extensions lifecycle. Supports $filter (eq).
      */
     @SerializedName(value = "status", alternate = {"Status"})
     @Expose
@@ -65,7 +65,7 @@ public class SchemaExtension extends Entity implements IJsonBackedObject {
 
     /**
      * The Target Types.
-     * Set of Microsoft Graph types (that can support extensions) that the schema extension can be applied to. Select from administrativeUnit, contact, device, event, group, message, organization, post, or user.
+     * Set of Microsoft Graph types (that can support extensions) that the schema extension can be applied to. Select from administrativeUnit, contact, device, event, group, message, organization, post, todoTask, todoTaskList, or user.
      */
     @SerializedName(value = "targetTypes", alternate = {"TargetTypes"})
     @Expose