+++ src/main/java/com/microsoft/graph/models/ServicePrincipal.java
@@ -22,6 +22,7 @@ import com.microsoft.graph.models.ClaimsMappingPolicy;
 import com.microsoft.graph.models.DirectoryObject;
 import com.microsoft.graph.models.DelegatedPermissionClassification;
 import com.microsoft.graph.models.Endpoint;
+import com.microsoft.graph.models.FederatedIdentityCredential;
 import com.microsoft.graph.models.HomeRealmDiscoveryPolicy;
 import com.microsoft.graph.models.OAuth2PermissionGrant;
 import com.microsoft.graph.models.TokenIssuancePolicy;
@@ -31,6 +32,7 @@ import com.microsoft.graph.requests.ClaimsMappingPolicyCollectionPage;
 import com.microsoft.graph.requests.DirectoryObjectCollectionPage;
 import com.microsoft.graph.requests.DelegatedPermissionClassificationCollectionPage;
 import com.microsoft.graph.requests.EndpointCollectionPage;
+import com.microsoft.graph.requests.FederatedIdentityCredentialCollectionPage;
 import com.microsoft.graph.requests.HomeRealmDiscoveryPolicyCollectionPage;
 import com.microsoft.graph.requests.OAuth2PermissionGrantCollectionPage;
 import com.microsoft.graph.requests.TokenIssuancePolicyCollectionPage;
@@ -390,6 +392,15 @@ public class ServicePrincipal extends DirectoryObject implements IJsonBackedObje
     public EndpointCollectionPage endpoints;
 
     /**
+     * The Federated Identity Credentials.
+     * 
+     */
+    @SerializedName(value = "federatedIdentityCredentials", alternate = {"FederatedIdentityCredentials"})
+    @Expose
+	@Nullable
+    public FederatedIdentityCredentialCollectionPage federatedIdentityCredentials;
+
+    /**
      * The Home Realm Discovery Policies.
      * The homeRealmDiscoveryPolicies assigned to this service principal. Supports $expand.
      */
@@ -477,6 +488,10 @@ public class ServicePrincipal extends DirectoryObject implements IJsonBackedObje
 
         if (json.has("endpoints")) {
             endpoints = serializer.deserializeObject(json.get("endpoints"), EndpointCollectionPage.class);
+        }
+
+        if (json.has("federatedIdentityCredentials")) {
+            federatedIdentityCredentials = serializer.deserializeObject(json.get("federatedIdentityCredentials"), FederatedIdentityCredentialCollectionPage.class);
         }
 
         if (json.has("homeRealmDiscoveryPolicies")) {