+++ src/main/java/com/microsoft/graph/requests/IdentityContainerRequestBuilder.java
@@ -8,7 +8,6 @@ package com.microsoft.graph.requests;
 import com.microsoft.graph.http.IRequestBuilder;
 import com.microsoft.graph.core.ClientException;
 import com.microsoft.graph.models.IdentityContainer;
-import com.microsoft.graph.requests.ConditionalAccessRootRequestBuilder;
 import com.microsoft.graph.requests.IdentityApiConnectorCollectionRequestBuilder;
 import com.microsoft.graph.requests.IdentityApiConnectorRequestBuilder;
 import com.microsoft.graph.requests.B2xIdentityUserFlowCollectionRequestBuilder;
@@ -17,6 +16,7 @@ import com.microsoft.graph.requests.IdentityProviderBaseCollectionRequestBuilder
 import com.microsoft.graph.requests.IdentityProviderBaseRequestBuilder;
 import com.microsoft.graph.requests.IdentityUserFlowAttributeCollectionRequestBuilder;
 import com.microsoft.graph.requests.IdentityUserFlowAttributeRequestBuilder;
+import com.microsoft.graph.requests.ConditionalAccessRootRequestBuilder;
 import java.util.Arrays;
 import java.util.EnumSet;
 import javax.annotation.Nullable;
@@ -65,16 +65,6 @@ public class IdentityContainerRequestBuilder extends BaseRequestBuilder<Identity
     }
 
 
-
-    /**
-     * Gets the request builder for ConditionalAccessRoot
-     *
-     * @return the ConditionalAccessRootRequestBuilder instance
-     */
-    @Nonnull
-    public ConditionalAccessRootRequestBuilder conditionalAccess() {
-        return new ConditionalAccessRootRequestBuilder(getRequestUrlWithAdditionalSegment("conditionalAccess"), getClient(), null);
-    }
     /**
      *  Gets a request builder for the IdentityApiConnector collection
      *
@@ -154,5 +144,15 @@ public class IdentityContainerRequestBuilder extends BaseRequestBuilder<Identity
     @Nonnull
     public IdentityUserFlowAttributeRequestBuilder userFlowAttributes(@Nonnull final String id) {
         return new IdentityUserFlowAttributeRequestBuilder(getRequestUrlWithAdditionalSegment("userFlowAttributes") + "/" + id, getClient(), null);
+    }
+
+    /**
+     * Gets the request builder for ConditionalAccessRoot
+     *
+     * @return the ConditionalAccessRootRequestBuilder instance
+     */
+    @Nonnull
+    public ConditionalAccessRootRequestBuilder conditionalAccess() {
+        return new ConditionalAccessRootRequestBuilder(getRequestUrlWithAdditionalSegment("conditionalAccess"), getClient(), null);
     }
 }