+++ src/main/java/com/microsoft/graph/requests/PolicyRootRequest.java
@@ -15,6 +15,7 @@ import com.microsoft.graph.requests.ActivityBasedTimeoutPolicyRequestBuilder;
 import com.microsoft.graph.requests.AuthorizationPolicyRequestBuilder;
 import com.microsoft.graph.requests.ClaimsMappingPolicyCollectionRequestBuilder;
 import com.microsoft.graph.requests.ClaimsMappingPolicyRequestBuilder;
+import com.microsoft.graph.requests.CrossTenantAccessPolicyRequestBuilder;
 import com.microsoft.graph.requests.HomeRealmDiscoveryPolicyCollectionRequestBuilder;
 import com.microsoft.graph.requests.HomeRealmDiscoveryPolicyRequestBuilder;
 import com.microsoft.graph.requests.PermissionGrantPolicyCollectionRequestBuilder;