+++ src/main/java/com/microsoft/graph/securitynamespace/models/CaseOperationStatus.java
@@ -1,0 +1,46 @@
+// Template Source: Enum.java.tt
+// ------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.  See License in the project root for license information.
+// ------------------------------------------------------------------------------
+
+package com.microsoft.graph.security.models;
+
+
+/**
+ * The Enum Case Operation Status.
+*/
+public enum CaseOperationStatus
+{
+    /**
+    * not Started
+    */
+    NOT_STARTED,
+    /**
+    * submission Failed
+    */
+    SUBMISSION_FAILED,
+    /**
+    * running
+    */
+    RUNNING,
+    /**
+    * succeeded
+    */
+    SUCCEEDED,
+    /**
+    * partially Succeeded
+    */
+    PARTIALLY_SUCCEEDED,
+    /**
+    * failed
+    */
+    FAILED,
+    /**
+    * unknown Future Value
+    */
+    UNKNOWN_FUTURE_VALUE,
+    /**
+    * For CaseOperationStatus values that were not expected from the service
+    */
+    UNEXPECTED_VALUE
+}